Update Assignment1.py
Ran pylint to check the coding standard. Added a docustring in line 2, simplified the docustring in line 18, used an f-string to quickly print the value instead of running through all the workers outputs in line 59. I removed the comment "# Added a docustring to explain the purpose, input and output of the function
#Technically it should return either no value or a worker will find the next value of pi." as this belongs here and also initially improved the pylint score and allowed the code to run smoothly. Considered using fsum instead of mpmath to improve precision and speed.