Update Assignment1.py
I have removed the comm.reduce work I did because the job took too long to complete for me to compare it to other jobs. The previous version found 15 significant figures of pi. However, as mp.dps has been increased to provide greater accuracy and less rounding, the result should be the same.