Update Assignment2.py
Attempting task 3, first writing out the vectors so that I can see their layout clearly. I think I can call on them based on their rank but unsure as of now. The i, j and k components essentially vary so I can pass this through the classes which I have made. I need to write code to instantiate objects from my classes, so perhaps I must def a new vector for each? Or set them as objects as I have, then call on a def of area to compute with them, as before?