Docustrings, variance, pylint, gaussian
First added docustrings to classes and functions, will improve later.

Then completed the variance function and improved the snake case here.

Pylint showed some whitespace, long lines, snake case and names being used twice so that was changed, also was unhappy with dimensions=[1,6] so changed it back to dimensions=1 but later called dimensions in a list, will try to plot the gaussian beyond 1D.
Completed the gaussian plot and edited some areas to make code flow better. Added print statements accordingly and added for dim in [1, 6]: to print gaussian in 1D and 6D but would like to plot this and then add errors as error bars later and extend this to 6D and then over all space.

Also cleaned up code and removed excess.