Pylint, removed redundant code, added functions to calculate size of 4D and 5D spaces correctly with errors, tried to ensure only one output gets printed
Pylint was responsible for most changes. Removed redundant code, fixed long lines and whitespace. Also tried to limit outputs so I specified the rank and indented it all. Wasn't getting the right values of a 4D and 5D hyperspace so added some functions and improved the error calculation. Tried to properly define the transform of the gaussian by substitution by calculating/mapping the x, y, z components of the gaussian and the transforming it into new x, y, z with the substitution applied. Printed values at each step to find the error in the calculation.