Removed excess rank==0 statements and de-indented code, removed error checking graphs, added a separate sub function and plot, fixed excess error code,
Kept getting zero values previously so had to fix this in the monte carlo section. Made sure to calculate mean and variance across processors when evaluating the monte carlo integral then reducing this to combine the value into one output. Checked this output too.  Added a sub_function to call from and made sure it was running properly, added a plot for this.  Commented out the error bars to work on other parts of the code for now. Changed 6D plot into a proper gaussian plot like we see in optics etc. Tried to scale it into a 2D image. Added a separate statement to plot the substitution of the gaussian over all space.